Fixed D2CB user documentation related to java key store.
The example of code provided in README.md was not clear enough for
customers. the key store password was hard coded to nestwave and this
was not clear that it was an dummy example value.

Code was changed to try to get the password from sensitive-config.ini.
Comments were added to ensure customers understood this clearly.